fix cart fragment
