+++ app/src/main/AndroidManifest.xml
@@ -37,7 +37,7 @@
          android:name=".SearchScreen.SearchDashboardActivity"
          android:exported="false" />
         <activity
-            android:name=".Activity.MainActivity"
+            android:name=".CartScreen.Cart"
             android:exported="false" />
         <activity
             android:name=".AddressScreen.AddAddress"
@@ -70,7 +70,7 @@
             android:name=".AddressScreen.PickAddress"
             android:exported="false" />
         <activity
-            android:name=".CartScreen.Cart"
+            android:name=".Activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
