--- app/src/main/res/layout/activity_cart_screen.xml
@@ -1,101 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".CartScreen.Cart"
-    android:orientation="vertical">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="20dp"
-        android:orientation="horizontal">
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:text="Giỏ hàng"
-            android:textSize="23sp"
-            android:fontFamily="@font/inter_semibold"/>
-    </LinearLayout>
-<LinearLayout
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal"
-    android:layout_marginStart="20dp">
-    <CheckBox
-        android:id="@+id/checkboxSelectAll"
-        android:layout_width="30dp"
-        android:layout_height="30dp"/>
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Chọn tất cả"
-        android:layout_marginStart="10dp"/>
-</LinearLayout>
-<LinearLayout
-    android:layout_width="match_parent"
-    android:layout_height="0dp"
-    android:layout_weight="1"
-    android:orientation="vertical">
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvCart"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="20dp" />
-    <TextView
-        android:id="@+id/txtHomepageHead3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toStartOf="@+id/btnViewAll03"
-        android:text="Sản phẩm liên quan"
-        android:textColor="@color/md_theme_onSurface"
-        android:textSize="27.5dp"
-        android:fontFamily="@font/inter_semibold"
-        app:layout_constraintHorizontal_chainStyle="spread_inside"/>
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rcvRelateProduct"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
-        android:paddingStart="20dp"
-        />
-</LinearLayout>
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:weightSum="10">
-        <LinearLayout
-            android:layout_width="250dp"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:orientation="vertical"
-            android:layout_weight="6">
-            <TextView
-                android:id="@+id/textView3"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Tổng thanh toán"
-                android:textSize="22sp" />
-            <TextView
-                android:id="@+id/totalPrice"
-                android:layout_width="400dp"
-                android:layout_height="wrap_content"
-                android:text="0"
-                android:textColor="@color/md_theme_error"
-                android:textSize="28sp" />
-        </LinearLayout>
-        <Button
-            android:id="@+id/btnMuangay"
-            android:layout_weight="4"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:text="Mua ngay"
-            android:fontFamily="@font/inter_semibold"
-            android:backgroundTint="@color/md_theme_onPrimaryFixedVariant"
-            android:layout_marginEnd="10dp"/>
-    </LinearLayout>
-</LinearLayout>
